requirements.txt
absl-py==0.12.0
affine==2.3.1
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
astor==0.8.1
async-timeout==3.0.1
attrs==20.3.0
basicsr==1.2.0+10018c6
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==8.0.4
click-plugins==1.1.1
cligj==0.7.2
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
dominate==2.6.0
easydict==1.9
einops==0.4.1
fastrlock==0.6
fsspec==2021.4.0
future==0.18.2
fvcore==0.1.5.post20210825
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
grpcio==1.37.0
h5py==3.1.0
Hydra==2.5
hydra-core==1.0.6
idna==2.10
idna-ssl==1.1.0
image-similarity-measures==0.3.5
imageio==2.9.0
importlib-metadata==4.0.1
importlib-resources==5.1.2
iopath==0.1.9
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lmdb==1.3.0
lpips==0.1.4
Markdown==3.3.4
matplotlib==3.3.4
multidict==5.1.0
natsort==7.1.1
networkx==2.5.1
numpy==1.16.0
oauthlib==3.1.0
omegaconf==2.0.6
opencv-python==4.5.1.48
packaging==20.9
pandas==1.1.5
phasepack==1.5
Pillow==8.2.0
pip==21.0.1
portalocker==2.3.2
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyFFTW==0.12.0
pynvml==8.0.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==22.0.3
rasterio==1.2.10
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.2.1
setproctitle==1.2.2
setuptools==52.0.0.post20210125
sewar==0.4.4
six==1.15.0
sklearn==0.0
snuggs==1.4.7
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4
tensorflow-gpu==1.12.0
termcolor==1.1.0
test-tube==0.7.5
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.9.3
timm==0.3.2
torch==1.8.1
torchfile==0.1.0
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
ttach==0.0.3
typing-extensions==3.7.4.3
urllib3==1.26.4
visdom==0.1.8.9
vit-pytorch==0.32.2
websocket-client==1.0.0
Werkzeug==1.0.1
wheel==0.36.2
yacs==0.1.8
yarl==1.6.3
zipp==3.4.1
